Fix hunspell header include path to be correct wrt pkgconf cflags
On every platform I checked include/hunspell is included in the cflags.
Thus, including <hunspell/hunspell.hxx> only works in the event that the top
level include dir is in the include path.  When using pkgsrc this isn't
necessarily the case.

This (the cflags including the hunspell directory specifically) is also true on
the other platforms I checked (Fedora and Debian) so this change should be safe.
On every platform I checked (Fedora, Debian, NetBSD) the hunspell cflags set up the include path so hunspell.hxx is already included