Add the process-end-of-output function to return the marker that is set at the end of the receieved date from a sub process.
update unix-process to use the new function and not turn off highlighting.
Put the key binding in a unix-process.key file, as is now the pattern.

Use EmacsBufferRef instead of a raw EmacsBuffer * that will lead to a crash
if the buffer is deleted.

make stuff_buffer into the ProcessChannelInput class as handleReceivedInput.