Sending signals with killpg is not doing what was expected. Until that is fixed send the ^C and ^\ chars for int and quit.
Apart from the shell named functions use pr- as a common prefix for the
process functions.

Use ~ prefixed local var names.

Remove the abbrev tricks. Its not that useful anymore.

rename to pr-kill-last-command-output that reflects what the function does.

Make pr-newline not set the mark as a side effect of its actions.